Add support for DualAnalog and DualShock controllers. Ape Escape is working fine in Dualshock mode. Also added a toggle for the 3 different controller modes.
Thanks James and jbd1986 for the additional input code and assistance !

+++ src/port/sdl/port.h
@@ -34,7 +34,7 @@ static inline int fsync(int f)
 unsigned get_ticks(void);
 void wait_ticks(unsigned s);
 void pad_update(void);
-unsigned short pad_read(int num);
+uint64_t pad_read(int num);
 
 void video_flip(void);
 #ifdef GPU_DFXVIDEO
